requirements.txt
# Requirements for vision_browser application

# For image processing
# must have tesseract installed
opencv-python>=4.5.0
pytesseract>=0.3.8

# For browser automation and interaction
selenium>=4.0.0
playwright>=1.30.0

# For AI and automation
openai>=1.0.0

# For environment variables
python-dotenv>=0.19.0

# Optional but recommended for working with arrays and data processing
numpy>=1.21.0
Pillow>=8.0.0  # Python Imaging Library for image processing

# Add other dependencies as needed 